moved character position function to utils
